revised RouteA.py and others
