+++ 2,3-butanediol synthesis/models/RouteB.py
@@ -4,159 +4,161 @@ import ode
 
 #enzyme setup
 
-GluGA3Pb_vmax =
-GluGA3Pb_rate =
-GluGA3Pb_km = 
-GluGA3Pb_gamma =
+GluGA3Pb_vmax = 0.0
+GluGA3Pb_rate = 4.14
+GluGA3Pb_km = 0.451
+GluGA3Pb_gamma = 0.0
 
-GA3PPvb_vmax =
-GA3PPvb_rate = 
-GA3PPvb_km = 
-GA3PPvb_gamma =
+#not confirmed yet
 
-PvAlphaALb_vmax =
-PvAlphaALb_rate =
-PvAlphaALb_km =
-PvAlphaALb_gamma =
+GA3PPvb_vmax = 0.51
+GA3PPvb_rate = 4.14
+GA3PPvb_km = 0.7
+GA3PPvb_gamma = 0.0
 
-AlphaAlDAb_vmax =
-AlphaAlDAb_rate =
-AlphaAlDAb_km =
-AlphaAlDAb_gamma =
+PvAlphaALb_vmax = 0
+PvAlphaALb_rate = 21.3
+PvAlphaALb_km = 0.92
+PvAlphaALb_gamma = 0.0
 
-DAATb_vmax = 
-DAATb_rate =
-DAATb_km =
-DAATb_gamma =
+AlphaALDAb_vmax = 0.9
+AlphaALDAb_rate = 21.5
+AlphaALDAb_km = 0.672
+AlphaALDAb_gamma = 0.0
 
-ATBDOb_vmax =
-ATBDOb_rate =
-ATBDOb_km =
-ATBDOb_gamma =
+DAATb_vmax = 0
+DAATb_rate = 59.1
+DAATb_km = 0.77
+DAATb_gamma = 0.0
+
+ATBDOb_vmax = 8.2
+ATBDOb_rate = 0.11
+ATBDOb_km = 0.732
+ATBDOb_gamma = 0.0
 
 
 #substrate setup
 
-Glub_vmax =
-Glub_gamma =
+Glub_vmax = 0
+Glub_gamma = 0
 
-GA3Pb_gamma =
-Pvb_gamma =
-AlphaALb_gamma =
-DAb_gamma =
-ATb_gamma =
-BDOb_gamma =
+GA3Pb_gamma = 0
+Pvb_gamma = 0
+AlphaALb_gamma = 0
+DAb_gamma = 0
+ATb_gamma = 0
+BDOb_gamma = 0
 
 
 #ODE system setup, reactant and product concentrations input respectively
 
 y = range(13)
 #[GluGA3Pb]
-y[0] = 
+y[0] = 4.07e-05
 #[GA3PPvb]
-y[1] =
-#[PvAlphaALb]
-y[2] =
+y[1] = 1.5e-05
+#[PvAlphaALb], i.e. acetolactate synthase
+y[2] = 0.03
 #[AlphaALDAb]
-y[3] =
+y[3] = 1e-06
 #[DAATb]
-y[4] =
+y[4] = 1e-06
 #[ATBDOb]
-y[5] =
+y[5] = 1.2e-06
 
 #[Glub]
-y[6] =
+y[6] = 1e-05
 #[GA3Pb]
-y[7] =
+y[7] = 0
 #[Pvb]
-y[8] =
+y[8] = 0
 #[AlphaALb]
-y[9] =
+y[9] = 0
 #[DAb]
-y[10] =
+y[10] = 0
 #[ATb]
-y[11] =
+y[11] = 0
 #[BDOb]
-y[12] =
+y[12] = 0
 
 #individual definition
 
 def GluGA3Pb(t, y):
 	production = GluGA3Pb_vmax
-	degradation = GluGA3Pb_gamma * [GluGA3Pb]
+	degradation = GluGA3Pb_gamma * y[0]
 	usage = 0.0 
 	return production - degradation - usage
 
 def GA3PPvb(t, y):
 	production = GA3PPvb_vmax
-	degradation = GA3PPvb_gamma * [GA3PPvb]
+	degradation = GA3PPvb_gamma * y[1]
 	usage = 0.0
 	return production - degradation - usage
 
 def PvAlphaALb(t, y):
 	production = PvAlphaALb_vmax
-	degradation = PvAlphaALb_gamma * [PvAlphaALb]
+	degradation = PvAlphaALb_gamma * y[2]
 	usage = 0.0
 	return production - degradation - usage
 	
 def AlphaALDAb(t, y):
 	production = AlphaALDAb_vmax
-	degradation = AlphaALDAb_gamma * [AlphaALDAb]
+	degradation = AlphaALDAb_gamma * y[3]
 	usage = 0.0
 	return production - degradation - usage
 	
 def DAATb(t, y):
 	production = DAATb_vmax
-	degradation = DAATb_gamma * [DAATb]
+	degradation = DAATb_gamma * y[4]
 	usage = 0.0
 	return production - degradation - usage
 	
 def ATBDOb(t, y):
 	production = ATBDOb_vmax
-	degradation = ATBDOb_gamma * [ATBDOb]
+	degradation = ATBDOb_gamma * y[5]
 	usage = 0.0
 	return production - degradation - usage
 	
 
 def Glub(t, y):
 	production = Glub_vmax
-	degradation = Glub_gamma * [Glub]
-	usage = ([Glub] * [GA3Pb] * GluGA3Pb_rate)/([Glub] + GluGA3Pb_km)
+	degradation = Glub_gamma * y[6]
+	usage = (y[6] * y[7] * GluGA3Pb_rate)/(y[6] + GluGA3Pb_km)
 	return production - degradation - usage
 	
 def GA3Pb(t, y):
-	production = ([Glub] * [GA3Pb] * GluGA3Pb_rate)/([Glub] + GluGA3Pb_km)
-	degradation = GA3Pb_gamma * [GA3Pb]
-	usage = ([GA3Pb] * [Pvb] * GA3PPvb_rate)/([GA3P] + GA3PPvb_km)
+	production = (y[6] * y[7] * GluGA3Pb_rate)/(y[6] + GluGA3Pb_km)
+	degradation = GA3Pb_gamma * y[7]
+	usage = (y[7] * y[8] * GA3PPvb_rate)/(y[7] + GA3PPvb_km)
 	return production - degradation - usage
 	
 def Pvb(t, y):
-	production = ([GA3Pb] * [Pvb] * GA3PPvb_rate)/([GA3P] + GA3PPvb_km)
-	degradation = Pvb_gamma * [Pvb]
-	usage = ([Pvb] * [AlphaALb] * PvbAlphaALb_rate)/([Pvb] + PvbAlphaALb_km)
+	production = (y[7] * y[8] * GA3PPvb_rate)/(y[7] + GA3PPvb_km)
+	degradation = Pvb_gamma * y[8]
+	usage = (y[8] * y[9] * PvAlphaALb_rate)/(y[8] + PvAlphaALb_km)
 	return production - degradation - usage
 	
 def AlphaALb(t, y):
-	production = ([Pvb] * [AlphaALb] * PvbAlphaALb_rate)/([Pvb] + PvbAlphaALb_km)
-	degradation = AlphaALb_gamma * [AlphaALb]
-	usage = ([AlphaALb] * [DAb] * AlphaALDAb_rate)/([AlphaALb] + AlphaALDAb_km)
+	production = (y[8] * y[9] * PvAlphaALb_rate)/(y[8] + PvAlphaALb_km)
+	degradation = AlphaALb_gamma * y[9]
+	usage = (y[9] * y[10] * AlphaALDAb_rate)/(y[9] + AlphaALDAb_km)
 	return production - degradation - usage
 	
 def DAb(t, y):
-	production = ([AlphaALb] * [DAb] * AlphaALDAb_rate)/([AlphaALb] + AlphaALDAb_km)
-	degradation = DAb_gamma * [DAb]
-	usage = ([DAb] * [ATb] * DAATb_rate)/([DAb] + DAATb_km)
+	production = (y[9] * y[10] * AlphaALDAb_rate)/(y[9] + AlphaALDAb_km)
+	degradation = DAb_gamma * y[10]
+	usage = (y[10] * y[11] * DAATb_rate)/(y[10] + DAATb_km)
 	return production - degradation - usage
 	
 def ATb(t, y):
-	production = ([DAb] * [ATb] * DAATb_rate)/([DAb] + DAATb_km)
-	degradation = ATb_gamma * [ATb]
-	usage = ([ATb] * [BDOb] * ATBDOb_rate)/([ATb] + ATBDOb_km)
+	production = (y[10] * y[11] * DAATb_rate)/(y[10] + DAATb_km)
+	degradation = ATb_gamma * y[11]
+	usage = (y[11] * y[12] * ATBDOb_rate)/(y[11] + ATBDOb_km)
 	return production - degradation - usage
 	
 def BDOb(t, y):
-	production = ([ATb] * [BDOb] * ATBDOb_rate)/([ATb] + ATBDOb_km)
-	degradation = BDOb_gamma * [BDOb]
+	production = (y[11] * y[12] * ATBDOb_rate)/(y[11] + ATBDOb_km)
+	degradation = BDOb_gamma * y[12]
 	usage = 0.0
 	return production - degradation - usage
